+++ packages/elements/src/components/ui/button-links/FlatButtonLink.tsx
@@ -1,0 +1,19 @@
+import cx from "classnames";
+import * as React from "react";
+import { PrimaryButtonLink, PrimaryButtonLinkProps } from "./PrimaryButtonLink";
+import styles from "../buttons/FlatButton.module.css";
+
+export interface FlatButtonLinkProps extends PrimaryButtonLinkProps {
+  inverted?: boolean;
+}
+
+export const FlatButtonLink: React.FC<FlatButtonLinkProps> = ({
+  className,
+  inverted,
+  ...props
+}) => (
+  <PrimaryButtonLink
+    className={cx(className, styles.flatButton, inverted && styles.inverted)}
+    {...props}
+  />
+);