+++ packages/elements/src/components/ui/buttons/StandardButton.tsx
@@ -1,3 +1,0 @@
-import { PrimaryButton } from "./PrimaryButton";
-
-export const StandardButton = PrimaryButton;