+++ packages/elements/src/components/ui/buttons/common/ButtonLabelFactory.ts
@@ -1,0 +1,15 @@
+export const getButtonLabel = (
+  label: string | undefined,
+  success: boolean,
+  successLabel: string | undefined,
+  loading: boolean,
+  loadingLabel: string | undefined
+): string | undefined => {
+  if (success) {
+    return successLabel;
+  } else if (loading) {
+    return loadingLabel;
+  } else {
+    return label;
+  }
+};