+++ packages/panels/src/index.ts
@@ -1,3 +1,5 @@
+export * from "./components/action-menu-button/ActionMenuButton";
+export * from "./components/checkbox-menu/CheckboxMenu";
 export * from "./components/nav-bar/NavBar";
 export * from "./components/nav-bar/NavBarButton";
 export * from "./components/nav-bar/NavBarMenuButton";