+++ CHANGELOG.md
@@ -1,5 +1,39 @@
 # Changelog
 
+## 2.1.0
+
+### Breaking changes
+
+- Button components no longer has `as` prop.
+
+To create links with button look, use `PrimaryButtonLink`, etc.
+
+### New components
+
+- `TabMenu` and `Tab`.
+
+`TabMenu` can have `Tab` elements as children, creating a tab menu bar.
+`Tab` have mostly the same props as buttons.
+
+- `CheckboxMenu` in `panels` package.
+
+A checkbox and a menu, with options to quick select checkboxes.
+
+- `ActionMenuButton` in `panels` package.
+
+A flat button with ellipses icon that opens an action menu.
+
+- `PrimaryButtonLink`
+- `FlatButtonLink`
+- `SecondaryButtonLink`
+
+They are `a` tags with same design and functionality as `PrimaryButton`, etc.
+
+### Improvements
+
+- Updated design of `ActionMenu`.
+- Fix issues in buttons where spinner was not aligned correctly.
+
 ## 2.0.9
 
 ### Bug fix